refactor: replace ContractAbi interface
