+++ src/common/components/pages/ContractDefinedSpace.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { cloneDeep, find, indexOf, isNil, mapValues, toString } from "lodash";
+import { find, indexOf, isNil, mapValues, toString } from "lodash";
 import React, { useCallback, useEffect, useMemo, useState } from "react";
 import { useAuthenticatorManager } from "@/authenticators/AuthenticatorManager";
 import { useAppStore } from "@/common/data/stores/app";
@@ -18,414 +18,9 @@ import { FaComments, FaDollarSign, FaLink } from "react-icons/fa6";
 import { FaExchangeAlt, FaStream } from "react-icons/fa";
 import { MobileContractDefinedSpace } from "./MobileSpace";
 import { BasescanResult } from "@/pages/api/basescan/contract";
+import { clankerTokenAbi } from "@/utils/clankerAbi";
 
 const FARCASTER_NOUNSPACE_AUTHENTICATOR_NAME = "farcaster:nounspace";
-
-// TODO: send to utils
-interface ContractAbi {
-  type: string;
-  name?: string;
-  inputs?: Array<{
-    name: string;
-    type: string;
-    indexed?: boolean;
-  }>;
-  outputs?: Array<{
-    name: string;
-    type: string;
-  }>;
-  stateMutability?: string;
-  constant?: boolean;
-  payable?: boolean;
-}
-
-export const clankerTokenAbi: ContractAbi[] = [
-  {
-    type: "constructor",
-    inputs: [
-      { name: "name_", type: "string" },
-      { name: "symbol_", type: "string" },
-      { name: "maxSupply_", type: "uint256" },
-      { name: "deployer_", type: "address" },
-      { name: "fid_", type: "uint256" },
-      { name: "image_", type: "string" },
-      { name: "castHash_", type: "string" },
-    ],
-    stateMutability: "nonpayable",
-  },
-  { type: "error", name: "CheckpointUnorderedInsertion" },
-  { type: "error", name: "ECDSAInvalidSignature" },
-  {
-    type: "error",
-    name: "ECDSAInvalidSignatureLength",
-    inputs: [{ name: "length", type: "uint256" }],
-  },
-  {
-    type: "error",
-    name: "ECDSAInvalidSignatureS",
-    inputs: [{ name: "s", type: "bytes32" }],
-  },
-  {
-    type: "error",
-    name: "ERC20ExceededSafeSupply",
-    inputs: [
-      { name: "increasedSupply", type: "uint256" },
-      { name: "cap", type: "uint256" },
-    ],
-  },
-  {
-    type: "error",
-    name: "ERC20InsufficientAllowance",
-    inputs: [
-      { name: "spender", type: "address" },
-      { name: "allowance", type: "uint256" },
-      { name: "needed", type: "uint256" },
-    ],
-  },
-  {
-    type: "error",
-    name: "ERC20InsufficientBalance",
-    inputs: [
-      { name: "sender", type: "address" },
-      { name: "balance", type: "uint256" },
-      { name: "needed", type: "uint256" },
-    ],
-  },
-  {
-    type: "error",
-    name: "ERC20InvalidApprover",
-    inputs: [{ name: "approver", type: "address" }],
-  },
-  {
-    type: "error",
-    name: "ERC20InvalidReceiver",
-    inputs: [{ name: "receiver", type: "address" }],
-  },
-  {
-    type: "error",
-    name: "ERC20InvalidSender",
-    inputs: [{ name: "sender", type: "address" }],
-  },
-  {
-    type: "error",
-    name: "ERC20InvalidSpender",
-    inputs: [{ name: "spender", type: "address" }],
-  },
-  {
-    type: "error",
-    name: "ERC2612ExpiredSignature",
-    inputs: [{ name: "deadline", type: "uint256" }],
-  },
-  {
-    type: "error",
-    name: "ERC2612InvalidSigner",
-    inputs: [
-      { name: "signer", type: "address" },
-      { name: "owner", type: "address" },
-    ],
-  },
-  {
-    type: "error",
-    name: "ERC5805FutureLookup",
-    inputs: [
-      { name: "timepoint", type: "uint256" },
-      { name: "clock", type: "uint48" },
-    ],
-  },
-  { type: "error", name: "ERC6372InconsistentClock" },
-  {
-    type: "error",
-    name: "InvalidAccountNonce",
-    inputs: [
-      { name: "account", type: "address" },
-      { name: "currentNonce", type: "uint256" },
-    ],
-  },
-  { type: "error", name: "InvalidShortString" },
-  { type: "error", name: "NotDeployer" },
-  {
-    type: "error",
-    name: "SafeCastOverflowedUintDowncast",
-    inputs: [
-      { name: "bits", type: "uint8" },
-      { name: "value", type: "uint256" },
-    ],
-  },
-  {
-    type: "error",
-    name: "StringTooLong",
-    inputs: [{ name: "str", type: "string" }],
-  },
-  {
-    type: "error",
-    name: "VotesExpiredSignature",
-    inputs: [{ name: "expiry", type: "uint256" }],
-  },
-  {
-    type: "event",
-    name: "Approval",
-    inputs: [
-      { indexed: true, name: "owner", type: "address" },
-      { indexed: true, name: "spender", type: "address" },
-      { indexed: false, name: "value", type: "uint256" },
-    ],
-  },
-  {
-    type: "event",
-    name: "DelegateChanged",
-    inputs: [
-      { indexed: true, name: "delegator", type: "address" },
-      { indexed: true, name: "fromDelegate", type: "address" },
-      { indexed: true, name: "toDelegate", type: "address" },
-    ],
-  },
-  {
-    type: "event",
-    name: "DelegateVotesChanged",
-    inputs: [
-      { indexed: true, name: "delegate", type: "address" },
-      { indexed: false, name: "previousVotes", type: "uint256" },
-      { indexed: false, name: "newVotes", type: "uint256" },
-    ],
-  },
-  { type: "event", name: "EIP712DomainChanged" },
-  {
-    type: "event",
-    name: "Transfer",
-    inputs: [
-      { indexed: true, name: "from", type: "address" },
-      { indexed: true, name: "to", type: "address" },
-      { indexed: false, name: "value", type: "uint256" },
-    ],
-  },
-  {
-    type: "function",
-    name: "CLOCK_MODE",
-    outputs: [{ name: "", type: "string" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "DOMAIN_SEPARATOR",
-    outputs: [{ name: "", type: "bytes32" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "allowance",
-    inputs: [
-      { name: "owner", type: "address" },
-      { name: "spender", type: "address" },
-    ],
-    outputs: [{ name: "", type: "uint256" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "approve",
-    inputs: [
-      { name: "spender", type: "address" },
-      { name: "value", type: "uint256" },
-    ],
-    outputs: [{ name: "", type: "bool" }],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "function",
-    name: "balanceOf",
-    inputs: [{ name: "account", type: "address" }],
-    outputs: [{ name: "", type: "uint256" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "burn",
-    inputs: [{ name: "value", type: "uint256" }],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "function",
-    name: "burnFrom",
-    inputs: [
-      { name: "account", type: "address" },
-      { name: "value", type: "uint256" },
-    ],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "function",
-    name: "castHash",
-    outputs: [{ name: "", type: "string" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "clock",
-    outputs: [{ name: "", type: "uint48" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "decimals",
-    outputs: [{ name: "", type: "uint8" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "delegate",
-    inputs: [{ name: "delegatee", type: "address" }],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "function",
-    name: "delegateBySig",
-    inputs: [
-      { name: "delegatee", type: "address" },
-      { name: "nonce", type: "uint256" },
-      { name: "expiry", type: "uint256" },
-      { name: "v", type: "uint8" },
-      { name: "r", type: "bytes32" },
-      { name: "s", type: "bytes32" },
-    ],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "function",
-    name: "delegates",
-    inputs: [{ name: "account", type: "address" }],
-    outputs: [{ name: "", type: "address" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "deployer",
-    outputs: [{ name: "", type: "address" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "eip712Domain",
-    outputs: [
-      { name: "fields", type: "bytes1" },
-      { name: "name", type: "string" },
-      { name: "version", type: "string" },
-      { name: "chainId", type: "uint256" },
-      { name: "verifyingContract", type: "address" },
-      { name: "salt", type: "bytes32" },
-      { name: "extensions", type: "uint256[]" },
-    ],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "fid",
-    outputs: [{ name: "", type: "uint256" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "getPastTotalSupply",
-    inputs: [{ name: "timepoint", type: "uint256" }],
-    outputs: [{ name: "", type: "uint256" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "getPastVotes",
-    inputs: [
-      { name: "account", type: "address" },
-      { name: "timepoint", type: "uint256" },
-    ],
-    outputs: [{ name: "", type: "uint256" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "getVotes",
-    inputs: [{ name: "account", type: "address" }],
-    outputs: [{ name: "", type: "uint256" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "image",
-    outputs: [{ name: "", type: "string" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "name",
-    outputs: [{ name: "", type: "string" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "nonces",
-    inputs: [{ name: "owner", type: "address" }],
-    outputs: [{ name: "", type: "uint256" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "numCheckpoints",
-    inputs: [{ name: "account", type: "address" }],
-    outputs: [{ name: "", type: "uint32" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "permit",
-    inputs: [
-      { name: "owner", type: "address" },
-      { name: "spender", type: "address" },
-      { name: "value", type: "uint256" },
-      { name: "deadline", type: "uint256" },
-      { name: "v", type: "uint8" },
-      { name: "r", type: "bytes32" },
-      { name: "s", type: "bytes32" },
-    ],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "function",
-    name: "symbol",
-    outputs: [{ name: "", type: "string" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "totalSupply",
-    outputs: [{ name: "", type: "uint256" }],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "transfer",
-    inputs: [
-      { name: "recipient", type: "address" },
-      { name: "amount", type: "uint256" },
-    ],
-    outputs: [{ name: "", type: "bool" }],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "function",
-    name: "transferFrom",
-    inputs: [
-      { name: "sender", type: "address" },
-      { name: "recipient", type: "address" },
-      { name: "amount", type: "uint256" },
-    ],
-    outputs: [{ name: "", type: "bool" }],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "function",
-    name: "version",
-    outputs: [{ name: "", type: "string" }],
-    stateMutability: "view",
-  },
-] as const;
 
 interface ContractDefinedSpaceProps {
   spaceId: string | null;
@@ -835,6 +430,7 @@ const ContractDefinedSpace = (props: ContractDefinedSpaceProps) => {
       );
       const contractData: BasescanResult = await contract.json();
       if (
+        contractData.contractFactory &&
         [
           "0x250c9FB2b411B48273f69879007803790A6AeA47".toLocaleLowerCase(), // v0
           "0x9B84fcE5Dcd9a38d2D01d5D72373F6b6b067c3e1".toLocaleLowerCase(), // v1
